requirements.txt
#Flask==3.0.0
#Flask_Cors==3.0.10
numpy==1.26.1
pandas==2.1.1
python-dotenv #==1.0.0
simpy #==4.0.2
SQLAlchemy #==2.0.22
pymysql #==1.1.0
dash
dash_bootstrap_components
plotly.express
dash_extensions==1.0.4
cryptography